requirements.txt
fastapi
requests
pycryptodome
uvicorn
